+++ package.json
@@ -71,7 +71,11 @@
     ]
   },
   "prettier": {
-    "singleQuote": true
+    "singleQuote": true,
+		"arrowParens": "always",
+		"semi": true,
+		"tabWidth": 2,
+		"printWidth": 80
   },
   "release": {
     "branches": "main",
